Re-applied Ciprian's NitPicks :)
